+++ inst/get_repos_tree_workflow.R
@@ -1,0 +1,13 @@
+devtools::load_all(".")
+
+test_gitstats <- create_gitstats() |>
+  set_github_host(
+    orgs = c("r-world-devs", "openpharma")
+  ) |>
+  set_gitlab_host(
+    orgs = c("mbtests", "mbtestapps")
+  )
+
+get_repos_trees(
+  gitstats = test_gitstats
+)